deprecated apis removed and refactor code
