+++ .gitignore
@@ -35,7 +35,7 @@ lerna-debug.log*
 !.vscode/extensions.json
 
 # env ignore
-.env
+# .env
 /build/Jenkinsfile
 env.backup
 env.backup.server
