--- src/credentials/credentials.controller.spec.ts
@@ -1,22 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { CredService } from 'src/services/cred/cred.service';
-import { SbrcService } from 'src/services/sbrc/sbrc.service';
-import { CredentialsController } from './credentials.controller';
-import { CredentialsService } from './credentials.service';
-
-describe('CredentialsController', () => {
-  let controller: CredentialsController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [CredentialsController],
-      providers: [CredentialsService, CredService, SbrcService]
-    }).compile();
-
-    controller = module.get<CredentialsController>(CredentialsController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
