--- src/credentials/credentials.service.spec.ts
@@ -1,20 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { CredService } from 'src/services/cred/cred.service';
-import { SbrcService } from 'src/services/sbrc/sbrc.service';
-import { CredentialsService } from './credentials.service';
-
-describe('CredentialsService', () => {
-  let service: CredentialsService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [CredentialsService, CredService, SbrcService],
-    }).compile();
-
-    service = module.get<CredentialsService>(CredentialsService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
