+++ src/credentials/credentials.service.ts
@@ -1,482 +1,50 @@
-//import { HttpService } from '@nestjs/axios';
-import { Injectable, InternalServerErrorException } from '@nestjs/common';
+import { Injectable, StreamableFile } from '@nestjs/common';
 import { HttpService } from '@nestjs/axios';
 import { AxiosRequestConfig } from 'axios';
-import axios from 'axios';
-import { it } from 'node:test';
-import { SingleCredentialDto } from './dto/singlecred-dto';
-import { BulkCredentialDto } from './dto/bulkCred-dto';
-import { Response } from 'express';
-import { CredService } from 'src/services/cred/cred.service';
-import { SbrcService } from 'src/services/sbrc/sbrc.service';
-import { TelemetryService } from 'src/services/telemetry/telemetry.service';
-import { KeycloakService } from 'src/services/keycloak/keycloak.service';
 
+//custom imports
+import jwt_decode from 'jwt-decode';
+import { createWriteStream, writeFile } from 'fs';
+import { Response, Request } from 'express';
+import * as wkhtmltopdf from 'wkhtmltopdf';
+import { AadharService } from '../services/aadhar/aadhar.service';
+import { SbrcService } from 'src/services/sbrc/sbrc.service';
+import { CredService } from 'src/services/cred/cred.service';
+import { KeycloakService } from 'src/services/keycloak/keycloak.service';
+import { TelemetryService } from 'src/services/telemetry/telemetry.service';
+const qr = require('qrcode');
+const { parse, HTMLElement } = require('node-html-parser');
+
+const path = require('path');
+import sharp from 'sharp';
+import { join } from 'path';
+import { log } from 'console';
+const zlib = require('zlib');
+const htmlMinifier = require('html-minifier');
+
+const jsdom = require('jsdom');
+const handlebars = require('handlebars');
+const fs = require('fs');
+const crypto = require('crypto');
+const jQuery = require('jquery');
+const cheerio = require('cheerio');
+
+const minimize = require('minimize');
 @Injectable()
 export class CredentialsService {
   constructor(
     private readonly httpService: HttpService,
+    private aadharService: AadharService,
     private credService: CredService,
     private sbrcService: SbrcService,
     private telemetryService: TelemetryService,
     private keycloakService: KeycloakService,
   ) { }
 
-  async issueBulkCredential(
-    credentialPlayload: BulkCredentialDto,
-    schemaId: string,
-    type: string,
-    response: Response,
-  ) {
-    console.log('credentialPlayload: ', credentialPlayload);
-    console.log('schemaId: ', schemaId);
-
-    var issuerId = '';
-
-    // find or create issuerId
-    //find udise in rc
-    let searchSchema = {
-      filters: {
-        udiseCode: {
-          eq: credentialPlayload.issuerDetail.udise,
-        },
-      },
-    };
-    //let searchSchoolDetail = await this.sbrcSearch(searchSchema, 'SchoolDetail')
-    let searchSchoolDetail = await this.sbrcService.sbrcSearch(
-      searchSchema,
-      'SchoolDetail',
-    );
-    console.log('searchSchoolDetail', searchSchoolDetail);
-
-    if (searchSchoolDetail.length > 0) {
-      issuerId = searchSchoolDetail[0].did;
-      console.log('issuerId', issuerId);
-    } else {
-      //let schoolDidRes = await this.generateDid(credentialPlayload.issuerDetail.udise)
-      let schoolDidRes = await this.credService.generateDid(
-        credentialPlayload.issuerDetail.udise,
-      );
-      console.log('schoolDidRes', schoolDidRes);
-
-      if (schoolDidRes) {
-        credentialPlayload.issuerDetail.schoolDid =
-          schoolDidRes[0].verificationMethod[0].controller;
-        //create schoolDetail in rc
-
-        let inviteSchema = {
-          schoolName: credentialPlayload.issuerDetail.schoolName,
-          udiseCode: credentialPlayload.issuerDetail.udise,
-          did: credentialPlayload.issuerDetail.schoolDid,
-        };
-        //let createSchoolDetail = await this.sbrcInvite(inviteSchema, 'SchoolDetail')
-        let createSchoolDetail = await this.sbrcService.sbrcInvite(
-          inviteSchema,
-          'SchoolDetail',
-        );
-        console.log('createSchoolDetail', createSchoolDetail);
-
-        if (createSchoolDetail) {
-          issuerId = credentialPlayload.issuerDetail.schoolDid;
-          console.log('issuerId', issuerId);
-        } else {
-          return response.status(200).send({
-            success: false,
-            status: 'sb_rc_register_error',
-            message: 'System Register Error ! Please try again.',
-            result: null,
-          });
-        }
-      } else {
-        return response.status(200).send({
-          success: false,
-          status: 'did_generate_error',
-          message: 'Identity Generation Failed ! Please Try Again.',
-          result: null,
-        });
-      }
-    }
-
-    //generate schema
-    //var schemaRes = await this.generateSchema(schemaId);
-    var schemaRes = await this.credService.generateSchema(schemaId);
-    console.log('schemaRes', schemaRes);
-
-    if (schemaRes) {
-      var responseArray = [];
-
-      // bulk import
-      for (const iterator of credentialPlayload.credentialSubject) {
-        if (credentialPlayload.credentialSubjectCommon.grade) {
-          iterator.grade = credentialPlayload.credentialSubjectCommon.grade;
-        }
-        if (credentialPlayload.credentialSubjectCommon.academic_year) {
-          iterator.academic_year =
-            credentialPlayload.credentialSubjectCommon.academic_year;
-        }
-        if (credentialPlayload.credentialSubjectCommon.benefitProvider) {
-          iterator.benefitProvider =
-            credentialPlayload.credentialSubjectCommon.benefitProvider;
-        }
-        if (credentialPlayload.credentialSubjectCommon.schemeName) {
-          iterator.schemeName =
-            credentialPlayload.credentialSubjectCommon.schemeName;
-        }
-        if (credentialPlayload.credentialSubjectCommon.schemeId) {
-          iterator.schemeId =
-            credentialPlayload.credentialSubjectCommon.schemeId;
-        }
-        if (credentialPlayload.credentialSubjectCommon.assessment) {
-          iterator.assessment =
-            credentialPlayload.credentialSubjectCommon.assessment;
-        }
-        if (credentialPlayload.credentialSubjectCommon.quarterlyAssessment) {
-          iterator.quarterlyAssessment =
-            credentialPlayload.credentialSubjectCommon.quarterlyAssessment;
-        }
-        if (credentialPlayload.credentialSubjectCommon.total) {
-          iterator.total = credentialPlayload.credentialSubjectCommon.total;
-        }
-        //list of schema update fields
-        if (credentialPlayload.issuerDetail.schoolName) {
-          iterator.school_name = credentialPlayload.issuerDetail.schoolName;
-        }
-        if (credentialPlayload.issuerDetail.udise) {
-          iterator.school_id = credentialPlayload.issuerDetail.udise;
-        }
-        if (credentialPlayload.credentialSubjectCommon.stateCode) {
-          iterator.stateCode =
-            credentialPlayload.credentialSubjectCommon.stateCode;
-        }
-        if (credentialPlayload.credentialSubjectCommon.stateName) {
-          iterator.stateName =
-            credentialPlayload.credentialSubjectCommon.stateName;
-        }
-        if (credentialPlayload.credentialSubjectCommon.districtCode) {
-          iterator.districtCode =
-            credentialPlayload.credentialSubjectCommon.districtCode;
-        }
-        if (credentialPlayload.credentialSubjectCommon.districtName) {
-          iterator.districtName =
-            credentialPlayload.credentialSubjectCommon.districtName;
-        }
-        if (credentialPlayload.credentialSubjectCommon.blockCode) {
-          iterator.blockCode =
-            credentialPlayload.credentialSubjectCommon.blockCode;
-        }
-        if (credentialPlayload.credentialSubjectCommon.blockName) {
-          iterator.blockName =
-            credentialPlayload.credentialSubjectCommon.blockName;
-        }
-
-        //generate did or find did
-        var aadhar_token = iterator.aadhar_token;
-
-        // find student
-        let name = iterator.student_name;
-        let dob = iterator.dob;
-        let searchSchema = {
-          filters: {
-            student_name: {
-              eq: name,
-            },
-            dob: {
-              eq: dob,
-            },
-          },
-        };
-        //const studentDetails = await this.sbrcSearch(searchSchema, 'StudentV2')
-        const studentDetails = await this.sbrcService.sbrcSearch(
-          searchSchema,
-          'StudentV2',
-        );
-        console.log('studentDetails', studentDetails);
-
-        if (studentDetails.length > 0) {
-          if (studentDetails[0]?.DID) {
-            iterator.id = studentDetails[0].DID;
-            let obj = {
-              issuerId: issuerId,
-              credSchema: schemaRes,
-              credentialSubject: iterator,
-              issuanceDate: credentialPlayload.vcData.issuanceDate,
-              expirationDate: credentialPlayload.vcData.expirationDate,
-            };
-            console.log('obj', obj);
-
-            //const cred = await this.issueCredentials(obj)
-            const cred = await this.credService.issueCredentials(obj);
-            //console.log("cred 34", cred)
-            if (cred) {
-              responseArray.push(cred);
-
-              //telemetry service called
-              this.telemetryService.telemetry({
-                id: iterator.id,
-                student_name: iterator.student_name,
-                dob: iterator.dob,
-                type: type,
-                result: 'credentials-issued',
-              });
-            } else {
-              responseArray.push({
-                student_name: iterator.student_name,
-                dob: iterator.dob,
-                error: 'unable to issue credentials!',
-              });
-              //telemetry service called
-              this.telemetryService.telemetry({
-                id: iterator.id,
-                student_name: iterator.student_name,
-                dob: iterator.dob,
-                type: type,
-                result: 'credentials-failed',
-              });
-            }
-          } else {
-            //let didRes = await this.generateDid(aadhar_token)
-            let didRes = await this.credService.generateDid(aadhar_token);
-
-            if (didRes) {
-              iterator.id = didRes[0].verificationMethod[0].controller;
-              //let updateRes = await this.sbrcUpdate({ DID: iterator.id }, 'StudentV2', studentDetails[0].osid)
-              let updateRes = await this.sbrcService.sbrcUpdate(
-                { DID: iterator.id },
-                'StudentV2',
-                studentDetails[0].osid,
-              );
-              if (updateRes) {
-                let obj = {
-                  issuerId: issuerId,
-                  credSchema: schemaRes,
-                  credentialSubject: iterator,
-                  issuanceDate: credentialPlayload.vcData.issuanceDate,
-                  expirationDate: credentialPlayload.vcData.expirationDate,
-                };
-                console.log('obj', obj);
-
-                if (iterator.id) {
-                  //const cred = await this.issueCredentials(obj)
-                  const cred = await this.credService.issueCredentials(obj);
-                  //console.log("cred 34", cred)
-                  if (cred) {
-                    responseArray.push(cred);
-                    //telemetry service called
-                    this.telemetryService.telemetry({
-                      id: iterator.id,
-                      student_name: iterator.student_name,
-                      dob: iterator.dob,
-                      type: type,
-                      result: 'credentials-issued',
-                    });
-                  } else {
-                    responseArray.push({
-                      student_name: iterator.student_name,
-                      dob: iterator.dob,
-                      error: 'unable to issue credentials!',
-                    });
-
-                    //telemetry service called
-                    this.telemetryService.telemetry({
-                      id: iterator.id,
-                      student_name: iterator.student_name,
-                      dob: iterator.dob,
-                      type: type,
-                      result: 'credentials-failed',
-                    });
-                  }
-                }
-              } else {
-                responseArray.push({
-                  student_name: iterator.student_name,
-                  dob: iterator.dob,
-                  error: 'unable to update did inside RC!',
-                });
-                //telemetry service called
-                this.telemetryService.telemetry({
-                  id: iterator.id,
-                  student_name: iterator.student_name,
-                  dob: iterator.dob,
-                  type: type,
-                  result: 'credentials-failed',
-                });
-              }
-            } else {
-              responseArray.push({
-                student_name: iterator.student_name,
-                dob: iterator.dob,
-                error: 'unable to generate student did!',
-              });
-
-              //telemetry service called
-              this.telemetryService.telemetry({
-                id: iterator.id,
-                student_name: iterator.student_name,
-                dob: iterator.dob,
-                type: type,
-                result: 'credentials-failed',
-              });
-            }
-          }
-        } else {
-          //let didRes = await this.generateDid(aadhar_token)
-          console.log('aadhar_token  205', aadhar_token);
-          if (aadhar_token) {
-            let didRes = await this.credService.generateDid(aadhar_token);
-
-            if (didRes) {
-              iterator.id = didRes[0].verificationMethod[0].controller;
-              let inviteSchema = {
-                student_id: iterator.student_id,
-                DID: iterator.id,
-                reference_id: iterator.reference_id,
-                aadhar_token: iterator.aadhar_token,
-                student_name: iterator.student_name,
-                dob: iterator.dob,
-                school_type: 'public',
-                meripehchan_id: '',
-                username: (
-                  iterator.student_name.split(' ')[0] +
-                  '@' +
-                  iterator.dob.split('/').join('')
-                ).toLowerCase(),
-                aadhaar_status: 'verified',
-                aadhaar_enc: '',
-                gender: iterator?.gender ? iterator.gender : '',
-                school_udise: iterator.school_id,
-                school_name: iterator.school_name,
-                stateCode: iterator.stateCode,
-                stateName: iterator.stateName,
-                districtCode: iterator.districtCode,
-                districtName: iterator.districtName,
-                blockCode: iterator.blockCode,
-                blockName: iterator.blockName,
-              };
-              console.log('inviteSchema', inviteSchema);
-              //let createStudent = await this.sbrcInvite(inviteSchema, 'StudentV2')
-              let createStudent = await this.sbrcService.sbrcInvite(
-                inviteSchema,
-                'StudentV2',
-              );
-              console.log('createStudent', createStudent);
-
-              if (createStudent) {
-                let obj = {
-                  issuerId: issuerId,
-                  credSchema: schemaRes,
-                  credentialSubject: iterator,
-                  issuanceDate: credentialPlayload.vcData.issuanceDate,
-                  expirationDate: credentialPlayload.vcData.expirationDate,
-                };
-                console.log('obj', obj);
-
-                //const cred = await this.issueCredentials(obj)
-                const cred = await this.credService.issueCredentials(obj);
-                //console.log("cred 34", cred)
-                if (cred) {
-                  responseArray.push(cred);
-                  //telemetry service called
-                  this.telemetryService.telemetry({
-                    id: iterator.id,
-                    student_name: iterator.student_name,
-                    dob: iterator.dob,
-                    type: type,
-                    result: 'credentials-issued',
-                  });
-                } else {
-                  responseArray.push({
-                    student_name: iterator.student_name,
-                    dob: iterator.dob,
-                    error: 'unable to issue credentials!',
-                  });
-
-                  //telemetry service called
-                  this.telemetryService.telemetry({
-                    id: iterator.id,
-                    student_name: iterator.student_name,
-                    dob: iterator.dob,
-                    type: type,
-                    result: 'credentials-failed',
-                  });
-                }
-              } else {
-                responseArray.push({
-                  student_name: iterator.student_name,
-                  dob: iterator.dob,
-                  error: 'unable to create student in RC!',
-                });
-
-                //telemetry service called
-                this.telemetryService.telemetry({
-                  id: iterator.id,
-                  student_name: iterator.student_name,
-                  dob: iterator.dob,
-                  type: type,
-                  result: 'credentials-failed',
-                });
-              }
-            } else {
-              responseArray.push({
-                student_name: iterator.student_name,
-                dob: iterator.dob,
-                error: 'unable to generate student did!',
-              });
-
-              //telemetry service called
-              this.telemetryService.telemetry({
-                id: iterator.id,
-                student_name: iterator.student_name,
-                dob: iterator.dob,
-                type: type,
-                result: 'credentials-failed',
-              });
-            }
-          } else {
-            responseArray.push({
-              student_name: iterator.student_name,
-              dob: iterator.dob,
-              error: 'aadhar_token not found!',
-            });
-
-            //telemetry service called
-            this.telemetryService.telemetry({
-              id: iterator.id,
-              student_name: iterator.student_name,
-              dob: iterator.dob,
-              type: type,
-              result: 'credentials-failed',
-            });
-          }
-        }
-      }
-
-      //bulk import response
-      console.log('responseArray.length', responseArray.length);
-      if (responseArray.length > 0) {
-        return response.status(200).send({
-          success: true,
-          status: 'student_cred_bulk_api_success',
-          message: 'Student Cred Bulk API Success.',
-          result: responseArray,
-        });
-      } else {
-        return response.status(200).send({
-          success: false,
-          status: 'did_cred_generate_error',
-          message:
-            'User Identity and Credentials Generation Failed. Try Again.',
-          result: null,
-        });
-      }
-    } else {
-      return response.status(200).send({
-        success: false,
-        status: 'did_cred_generate_error',
-        message: 'User Identity and Credentials Generation Failed. Try Again.',
-        result: null,
-      });
-    }
-  }
+  //axios call
+  md5 = require('md5');
+  moment = require('moment');
+  qs = require('qs');
 
   async getSchema(id: string, response: Response) {
     console.log('id: 172', id);
@@ -497,169 +65,6 @@ export class CredentialsService {
         success: false,
         status: 'cred_schema_api_failed',
         message: 'Credentials Schema Failed ! Please Try Again.',
-        result: null,
-      });
-    }
-  }
-
-  async issueSingleCredential(
-    credentialPlayload: SingleCredentialDto,
-    schemaId: string,
-    response: Response,
-  ) {
-    console.log('credentialPlayload: ', credentialPlayload);
-    console.log('schemaId: ', schemaId);
-
-    var payload = credentialPlayload;
-
-    var osid = payload.credentialSubject.osid;
-    var student_osid = payload.credentialSubject.student_osid;
-
-    //var issuerId = "did:ulp:f08f7782-0d09-4c47-aacb-9092113bc33e"
-    var issuerId = credentialPlayload.issuer;
-    console.log('issuerId', issuerId);
-
-    //generate schema
-    console.log('schemaId', schemaId);
-
-    //const schemaRes = await this.generateSchema(schemaId);
-    const schemaRes = await this.credService.generateSchema(schemaId);
-
-    console.log('schemaRes', schemaRes);
-
-    if (schemaRes) {
-      let aadhar_token = payload.credentialSubject.aadhar_token;
-      console.log('aadhar_token', aadhar_token);
-      //const didRes = await this.generateDid(aadhar_token);
-      const didRes = await this.credService.generateDid(aadhar_token);
-      console.log('didRes 191', didRes);
-
-      if (!didRes?.error) {
-        var did = didRes[0].verificationMethod[0].controller;
-        payload.credentialSubject.id = did;
-        delete payload.credentialSubject.osid;
-        delete payload.credentialSubject.student_osid;
-        //fix for enroll on 4 may
-        let enrolled_on = payload.credentialSubject.enrollon;
-        delete payload.credentialSubject.enrollon;
-        payload.credentialSubject.enrolled_on = enrolled_on;
-
-        let obj = {
-          issuerId: issuerId,
-          credSchema: schemaRes,
-          credentialSubject: payload.credentialSubject,
-          issuanceDate: payload.vcData.issuanceDate,
-          expirationDate: payload.vcData.expirationDate,
-        };
-        //console.log('obj', obj);
-        //const cred = await this.issueCredentials(obj)
-        //const cred = await this.credService.issueCredentials(obj)
-        //if (cred) {
-        //update did inside sbrc
-
-        //let updateStudentDetail = await this.sbrcUpdate({ "claim_status": "issued" }, 'StudentDetailV2', osid)
-        let updateStudentDetail = await this.sbrcService.sbrcUpdateEL(
-          { claim_status: 'issued' },
-          'StudentDetailV2',
-          osid,
-        );
-        if (updateStudentDetail?.error) {
-          return response.status(400).send({
-            success: false,
-            status: 'sb_rc_update_error',
-            message: 'System Update Error ! Please try again.',
-            result: updateStudentDetail?.error,
-          });
-        }
-        //console.log('updateStudentDetail', updateStudentDetail);
-        else {
-          //let updateStudent = await this.sbrcUpdate({ DID: did }, 'StudentV2', student_osid)
-          let updateStudent = await this.sbrcService.sbrcUpdateEL(
-            { DID: did },
-            'StudentV2',
-            student_osid,
-          );
-          //console.log('updateStudent', updateStudent);
-
-          if (updateStudent?.error) {
-            return response.status(400).send({
-              success: false,
-              status: 'sb_rc_update_error',
-              message: 'System Update Error ! Please try again.',
-              result: updateStudent?.error,
-            });
-          } else {
-            const cred = await this.credService.issueCredentials(obj);
-
-            if (cred) {
-              return response.status(200).send({
-                success: true,
-                status: 'Success',
-                message: 'Credentials generated successfully!',
-                result: cred,
-              });
-            } else {
-              return response.status(200).send({
-                success: false,
-                status: 'cred_issue_api_success',
-                message: 'Cred Issue API Success',
-                result: null,
-              });
-            }
-          }
-        }
-        // } else {
-        //     return response.status(200).send({
-        //         success: false,
-        //         status: 'cred_issue_api_success',
-        //         message: 'Cred Issue API Success',
-        //         result: null
-        //     })
-        // }
-      } else {
-        return response.status(200).send({
-          success: false,
-          status: 'did_generate_error',
-          message: 'Identity Generation Failed ! Please Try Again.',
-          result: null,
-        });
-      }
-    } else {
-      return response.status(200).send({
-        success: false,
-        status: 'did_cred_generate_error',
-        message: 'User Identity and Credentials Generation Failed. Try Again.',
-        result: null,
-      });
-    }
-  }
-
-  async rejectStudent(
-    credentialPlayload: SingleCredentialDto,
-    response: Response,
-  ) {
-    console.log('rejectStudentV2');
-    var payload = credentialPlayload;
-    var osid = payload.credentialSubject.osid;
-    //let updateRes = await this.sbrcUpdate({ "claim_status": "rejected" }, 'StudentDetailV2', osid)
-    let updateRes = await this.sbrcService.sbrcUpdate(
-      { claim_status: 'rejected' },
-      'StudentDetailV2',
-      osid,
-    );
-    console.log('updateRes 313', updateRes);
-    if (updateRes) {
-      return response.status(200).send({
-        success: true,
-        status: 'sb_rc_update_success',
-        message: 'System Update Success',
-        result: null,
-      });
-    } else {
-      return response.status(200).send({
-        success: false,
-        status: 'sb_rc_update_error',
-        message: 'System Update Error ! Please try again.',
         result: null,
       });
     }
@@ -724,261 +129,491 @@ export class CredentialsService {
     }
   }
 
-  //Verify credentials
-  public async verifyCertificate(payload, response) {
-    let id = 'did:ulp:c65940d4-c58d-450b-b37a-b2c7cf57b139' // valid id
-    //let id = 'did:ulp:c65940d4-c58d-450b-b37a-b2c7cf57b134' // invalid id
-    let config = {
-      method: 'get',
-      maxBodyLength: Infinity,
-      url: `${process.env.CRED_URL}/credentials/${id}/verify`,
-      headers: {}
-
-    };
-    try {
-      const res = await axios.request(config)
-      console.log("res", res.data)
-      if (res) {
-        return response.status(200).json(
-          {
-            "verified": true,
-            "results": [
-              {
-                "proof": {
-                  "@context": "https://w3id.org/security/v2",
-                  "type": "RsaSignature2018",
-                  "created": "2023-06-08T12:15:13Z",
-                  "verificationMethod": "did:india",
-                  "proofPurpose": "assertionMethod",
-                  "jws": "eyJhbGciOiJQUzI1NiIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..cgFAY8DuOVhZfNpcpDrQBRrkHjEPBh_i5kiRMoGCvgNFA_O7pF40ZBO_rk6lwiXBwBWWTUJEtLMaeGM8OKD5lCjL9RJ5dQFAI60qtEmbCL7sCqGYF6vXu_ryhhNiSZK4e07xW9rXXXzAH0ytpTYKxzG2T6qwbwUaozXxXw-GTvwZLmYtC8TlROWJxnPHCeETh_GLkTqOe_C4sqJgfsqsiGqNODKJpMjPBERR1z-GBe5X9sWUBrufuJiMPwjCocnwZfWAeyz78Xn-2fQf8I4iw3k9ytL9f_tpnnzR1WtIeLfbjga6XxGacNcfMHsBB9ETYMd0Bdum0hIDGVMPd2ygRw"
-                },
-                "verified": true,
-                "purposeResult": {
-                  "valid": true,
-                  "controller": {
-                    "@context": "https://w3id.org/security/v2",
-                    "id": "did:india",
-                    "publicKey": [
-                      {
-                        "@context": "https://w3id.org/security/v2",
-                        "id": "did:india",
-                        "type": "RsaVerificationKey2018",
-                        "controller": "https://sunbird.org/",
-                        "publicKeyPem": "-----BEGIN PUBLIC KEY-----\nMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAnXQalrgztecTpc+INjRQ8s73FSE1kU5QSlwBdICCVJBUKiuQUt7s+Z5epgCvLVAOCbP1mm5lV7bfgV/iYWDio7lzX4MlJwDedWLiufr3Ajq+79CQiqPaIbZTo0i13zijKtX7wgxQ78wT/HkJRLkFpmGeK3za21tEfttytkhmJYlwaDTEc+Kx3RJqVhVh/dfwJGeuV4Xc/e2NH++ht0ENGuTk44KpQ+pwQVqtW7lmbDZQJoOJ7HYmmoKGJ0qt2hrj15uwcD1WEYfY5N7N0ArTzPgctExtZFDmituLGzuAZfv2AZZ9/7Y+igshzfB0reIFdUKw3cdVTzfv5FNrIqN5pwIDAQAB\n-----END PUBLIC KEY-----\n"
-                      }
-                    ],
-                    "assertionMethod": [
-                      "did:india"
-                    ]
-                  }
-                }
-              }
-            ]
-          }
-        )
-      } else {
-        console.log("else", res.data)
-        return response.status(200).json({
-          "verified": false,
-          "results": []
+  //credentialsSearch
+  async credentialsSearch(
+    token: string,
+    type: string,
+    requestbody: any,
+    response: Response,
+  ) {
+    if (token && requestbody) {
+      const studentUsername = await this.keycloakService.verifyUserToken(token);
+      if (studentUsername?.error) {
+        return response.status(401).send({
+          success: false,
+          status: 'keycloak_token_bad_request',
+          message: 'You do not have access for this request.',
+          result: null,
         });
+      } else if (!studentUsername?.preferred_username) {
+        return response.status(400).send({
+          success: false,
+          status: 'keycloak_token_error',
+          message: 'Your Login Session Expired.',
+          result: null,
+        });
+      } else {
+        var data = JSON.stringify(requestbody);
+
+        const url = process.env.CRED_URL + '/credentials/search';
+
+        const config: AxiosRequestConfig = {
+          headers: {
+            'Content-Type': 'application/json',
+          },
+        };
+
+        let render_response = null;
+        try {
+          const observable = this.httpService.post(url, data, config);
+          const promise = observable.toPromise();
+          const response = await promise;
+          //console.log(JSON.stringify(response.data));
+          render_response = response.data;
+        } catch (e) {
+          //console.log(e);
+          render_response = { error: e };
+        }
+
+        if (render_response?.error) {
+          return response.status(400).send({
+            success: false,
+            status: 'cred_search_api_failed',
+            message: 'Credentials Search Failed ! Please Try Again.',
+            result: render_response,
+          });
+        } else {
+          let render_response_student = [];
+          /*if (type === 'student') {
+            for (let i = 0; i < render_response.length; i++) {
+              if (render_response[i]?.credentialSubject?.student_name) {
+                render_response_student.push(render_response[i]);
+              }
+            }
+          } else if (type === 'teacher') {
+            for (let i = 0; i < render_response.length; i++) {
+              if (!render_response[i]?.credentialSubject?.student_name) {
+                render_response_student.push(render_response[i]);
+              }
+            }
+          } else {
+            render_response_student = render_response;
+          }*/
+          render_response_student = render_response;
+          return response.status(200).send({
+            success: true,
+            status: 'cred_search_api_success',
+            message: 'Cred Search API Success',
+            result: render_response_student,
+          });
+        }
       }
-    } catch (err) {
-      //console.log("error", err)
-      return response.status(200).json({
-        "verified": false,
-        "results": []
+    } else {
+      return response.status(400).send({
+        success: false,
+        status: 'invalid_request',
+        message: 'Invalid Request. Not received token or requestbody.',
+        result: null,
       });
     }
-
-
   }
 
-  // //helper function
+  //credentialsSchema
+  async credentialsSchema(id: string, response: Response) {
+    if (id) {
+      const url = process.env.CRED_URL + '/credentials/schema/' + id;
 
-  // // credentials
-  // // generate schema
-  // async generateSchema(schemaId) {
-  //     var config = {
-  //         method: 'get',
-  //         maxBodyLength: Infinity,
-  //         url: `${process.env.SCHEMA_URL}/schema/jsonld?id=${schemaId}`,
-  //         headers: {}
-  //     };
+      var config = {
+        headers: {},
+      };
+      let response_text = null;
 
-  //     try {
-  //         const response = await axios(config)
-  //         console.log("response schema", response.data)
-  //         return response.data;
-  //     } catch (error) {
-  //         console.log("error schema", error)
-  //     }
-  // }
+      try {
+        const observable = this.httpService.get(url, config);
+        const promise = observable.toPromise();
+        const response = await promise;
+        //console.log(JSON.stringify(response.data));
+        response_text = response.data;
+      } catch (error) {
+        //console.log(e);
+        response_text = { error: error };
+      }
+      if (response_text?.error) {
+        return response.status(400).send({
+          success: false,
+          status: 'cred_schema_api_failed',
+          message: 'Credentials Schema Failed ! Please Try Again.',
+          result: response_text,
+        });
+      } else {
+        return response.status(200).send({
+          success: true,
+          status: 'cred_schema_api_success',
+          message: 'Cred Schema API Success',
+          result: response_text,
+        });
+      }
+    } else {
+      return response.status(400).send({
+        success: false,
+        status: 'invalid_request',
+        message: 'Invalid Request. Not received All Parameters.',
+        result: null,
+      });
+    }
+  }
 
-  // //generate did
-  // async generateDid(studentId) {
-  //     var data = JSON.stringify({
-  //         "content": [
-  //             {
-  //                 "alsoKnownAs": [
-  //                     `did.${studentId}`
-  //                 ],
-  //                 "services": [
-  //                     {
-  //                         "id": "IdentityHub",
-  //                         "type": "IdentityHub",
-  //                         "serviceEndpoint": {
-  //                             "@context": "schema.identity.foundation/hub",
-  //                             "@type": "UserServiceEndpoint",
-  //                             "instance": [
-  //                                 "did:test:hub.id"
-  //                             ]
-  //                         }
-  //                     }
-  //                 ]
-  //             }
-  //         ]
-  //     });
+  //renderTemplateSchema
+  async renderTemplateSchema(id: string, response: Response) {
+    if (id) {
+      const url = process.env.SCHEMA_URL + '/rendering-template/' + id;
 
-  //     var config = {
-  //         method: 'post',
-  //         maxBodyLength: Infinity,
-  //         url: `${process.env.DID_URL}/did/generate`,
-  //         headers: {
-  //             'Content-Type': 'application/json'
-  //         },
-  //         data: data
-  //     };
+      var config = {
+        headers: {},
+      };
+      let response_text = null;
 
-  //     try {
-  //         const response = await axios(config)
-  //         console.log("response did", response.data)
-  //         return response.data;
-  //     } catch (error) {
-  //         console.log("did error", error.message)
-  //     }
-  // }
+      try {
+        const observable = this.httpService.get(url, config);
+        const promise = observable.toPromise();
+        const response = await promise;
+        //console.log(JSON.stringify(response.data));
+        response_text = response.data;
+      } catch (error) {
+        //console.log(e);
+        //response_text = { error: error };
+      }
+      if (response_text == null) {
+        return response.status(400).send({
+          success: false,
+          status: 'render_template_schema_api_failed',
+          message: 'Render Template Schema Failed ! Please Try Again.',
+          result: null,
+        });
+      } else {
+        return response.status(200).send({
+          success: true,
+          status: 'render_template_schema_api_success',
+          message: 'Render Template Schema API Success',
+          result: response_text,
+        });
+      }
+    } else {
+      return response.status(400).send({
+        success: false,
+        status: 'invalid_request',
+        message: 'Invalid Request. Not received All Parameters.',
+        result: null,
+      });
+    }
+  }
 
-  // //issue cred
-  // async issueCredentials(payload) {
+  //credentialsSchemaJSON
+  async credentialsSchemaJSON(id: string, response: Response) {
+    if (id) {
+      const url = process.env.SCHEMA_URL + '/schema/jsonld?id=' + id;
 
-  //     var data = JSON.stringify({
-  //         "credential": {
-  //             "@context": [
-  //                 "https://www.w3.org/2018/credentials/v1",
-  //                 "https://www.w3.org/2018/credentials/examples/v1"
-  //             ],
-  //             "id": "did:ulp:b4a191af-d86e-453c-9d0e-dd4771067235",
-  //             "type": [
-  //                 "VerifiableCredential",
-  //                 "UniversityDegreeCredential"
-  //             ],
-  //             "issuer": `${payload.issuerId}`,
-  //             "issuanceDate": payload.issuanceDate,
-  //             "expirationDate": payload.expirationDate,
-  //             "credentialSubject": payload.credentialSubject,
-  //             "options": {
-  //                 "created": "2020-04-02T18:48:36Z",
-  //                 "credentialStatus": {
-  //                     "type": "RevocationList2020Status"
-  //                 }
-  //             }
-  //         },
-  //         "credentialSchemaId": payload.credSchema.id,
-  //         "tags": [
-  //             "tag1",
-  //             "tag2",
-  //             "tag3"
-  //         ]
-  //     });
+      const config: AxiosRequestConfig = {
+        headers: {},
+      };
+      let response_text = null;
 
-  //     var config = {
-  //         method: 'post',
-  //         maxBodyLength: Infinity,
-  //         url: 'http://64.227.185.154:3002/credentials/issue',
-  //         headers: {
-  //             'Content-Type': 'application/json'
-  //         },
-  //         data: data
-  //     };
+      try {
+        const observable = this.httpService.get(url, config);
+        const promise = observable.toPromise();
+        const response = await promise;
+        //console.log(JSON.stringify(response.data));
+        response_text = response.data;
+      } catch (error) {
+        //console.log(e);
+        response_text = { error: error };
+      }
+      if (response_text?.error) {
+        return response.status(400).send({
+          success: false,
+          status: 'cred_schema_json_api_failed',
+          message: 'Credentials Schema JSON Failed ! Please Try Again.',
+          result: response_text,
+        });
+      } else {
+        return response.status(200).send({
+          success: true,
+          status: 'cred_schema_json_api_success',
+          message: 'Cred Schema JSON API Success',
+          result: response_text,
+        });
+      }
+    } else {
+      return response.status(400).send({
+        success: false,
+        status: 'invalid_request',
+        message: 'Invalid Request. Not received All Parameters.',
+        result: null,
+      });
+    }
+  }
 
-  //     try {
+  //renderCredentials
+  async renderCredentials(
+    token: string,
+    requestbody: any,
+  ): Promise<string | StreamableFile> {
+    if (token) {
+      const studentUsername = await this.keycloakService.verifyUserToken(token);
+      if (studentUsername?.error) {
+        return 'Keycloak Student Token Expired';
+      } else if (!studentUsername?.preferred_username) {
+        return 'Keycloak Student Token Expired';
+      } else {
+        var data = JSON.stringify(requestbody);
 
-  //         const response = await axios(config)
-  //         console.log("cred response")
-  //         return response.data;
+        const url = process.env.CRED_URL + '/credentials/render';
 
-  //     } catch (e) {
-  //         console.log("cred error", e.message)
-  //     }
+        const config: AxiosRequestConfig = {
+          headers: {
+            'Content-Type': 'application/json',
+          },
+        };
 
-  // }
+        let render_response = null;
 
-  // // sbrc
-  // //invite
-  // async sbrcInvite(inviteSchema, entityName) {
-  //     let data = JSON.stringify(inviteSchema);
+        let certificateId = requestbody?.credential?.id;
 
-  //     let config = {
-  //         method: 'post',
-  //         url: process.env.REGISTRY_URL + 'api/v1/' + entityName + '/invite',
-  //         headers: {
-  //             'content-type': 'application/json',
-  //         },
-  //         data: data,
-  //     };
+        try {
+          const observable = this.httpService.post(url, data, config);
+          const promise = observable.toPromise();
+          const response = await promise;
 
-  //     try {
-  //         const response = await axios(config);
-  //         return response.data;
-  //     } catch (err) {
-  //         console.log("sb_rc_create err")
-  //     }
-  // }
+          render_response = response.data;
+        } catch (e) {
+          //console.log(e);
+          //render_response = { error: e };
+        }
+        if (render_response == null) {
+          return 'Credentials Render Failed ! Please Try Again.';
+        } else {
+          //return render_response;
 
-  // //search
-  // async sbrcSearch(searchSchema, entityName) {
-  //     let data = JSON.stringify(searchSchema);
+          try {
+            const url = process.env.VERIFICATION_URL + certificateId;
 
-  //     let config = {
-  //         method: 'post',
-  //         //url: process.env.REGISTRY_URL + 'api/v1/StudentV2/search',
-  //         url: process.env.REGISTRY_URL + 'api/v1/' + entityName + '/search',
-  //         headers: {
-  //             'Content-Type': 'application/json',
-  //         },
-  //         data: data,
-  //     };
-  //     try {
-  //         const response = await axios(config);
-  //         return response.data;
-  //     } catch (err) {
-  //         console.log("sb_rc_search err")
-  //     }
+            let modifiedHtml = null;
 
-  // }
+            const modified = await new Promise((resolve, reject) => {
+              qr.toDataURL(url, function (err, code) {
+                if (err) {
+                  resolve(null);
+                  return;
+                }
 
-  // //update
-  // async sbrcUpdate(updateSchema, entityName, osid) {
-  //     console.log("updateSchema", updateSchema)
-  //     console.log("entityName", entityName)
-  //     console.log("osid", osid)
-  //     let data = JSON.stringify(updateSchema);
+                if (code) {
+                  const newHtml = code;
 
-  //     let config = {
-  //         method: 'put',
-  //         url: process.env.REGISTRY_URL + 'api/v1/' + entityName + '/' + osid,
-  //         headers: {
-  //             'content-type': 'application/json',
-  //         },
-  //         data: data,
-  //     };
+                  const root = parse(render_response);
 
-  //     try {
-  //         const response = await axios(config);
-  //         return response.data;
-  //     } catch (err) {
-  //         console.log("sb_rc_update err")
-  //     }
-  // }
+                  // Find the img tag with id "qrcode"
+                  const qrcodeImg = root.querySelector('#qrcode');
+
+                  if (qrcodeImg) {
+                    qrcodeImg.setAttribute('src', newHtml);
+                    modifiedHtml = root.toString();
+
+                    resolve(modifiedHtml);
+                  } else {
+                    resolve(null);
+                  }
+                } else {
+                  resolve(null);
+                }
+              });
+            });
+
+            if (!modified) {
+              return null;
+            }
+
+            return new StreamableFile(
+              await wkhtmltopdf(modifiedHtml, {
+                pageSize: 'A4',
+                disableExternalLinks: true,
+                disableInternalLinks: true,
+                disableJavascript: true,
+              }),
+            );
+          } catch (e) {
+            //console.log(e);
+
+            return 'HTML to PDF Convert Fail';
+          }
+        }
+      }
+    } else {
+      return 'Student Token Not Received';
+    }
+  }
+
+  //renderCredentialsHTML
+  async renderCredentialsHTML(
+    token: string,
+    requestbody: any,
+    response: Response,
+  ) {
+    if (token) {
+      const studentUsername = await this.keycloakService.verifyUserToken(token);
+      if (studentUsername?.error) {
+        return response.status(401).send({
+          success: false,
+          status: 'keycloak_token_bad_request',
+          message: 'You do not have access for this request.',
+          result: null,
+        });
+      } else if (!studentUsername?.preferred_username) {
+        return response.status(400).send({
+          success: false,
+          status: 'keycloak_token_error',
+          message: 'Your Login Session Expired.',
+          result: null,
+        });
+      } else {
+        var data = JSON.stringify(requestbody);
+
+        const url = process.env.CRED_URL + '/credentials/render';
+
+        const config: AxiosRequestConfig = {
+          headers: {
+            'Content-Type': 'application/json',
+          },
+        };
+
+        let render_response = null;
+        try {
+          const observable = this.httpService.post(url, data, config);
+          const promise = observable.toPromise();
+          const response = await promise;
+          //console.log(JSON.stringify(response.data));
+          render_response = response.data;
+        } catch (e) {
+          //console.log(e);
+          //render_response = { error: e };
+        }
+
+        if (render_response == null) {
+          return response.status(400).send({
+            success: false,
+            status: 'render_api_failed',
+            message: 'Credentials Render Failed ! Please Try Again.',
+            result: null,
+          });
+        } else {
+          return response.status(200).send({
+            success: true,
+            status: 'render_api_success',
+            message: 'Cred Render API Success',
+            result: render_response,
+          });
+        }
+      }
+    } else {
+      return response.status(400).send({
+        success: false,
+        status: 'invalid_request',
+        message: 'Invalid Request. Not received token.',
+        result: null,
+      });
+    }
+  }
+
+  //credentialsVerify
+  async credentialsVerify(id: string, response: Response) {
+    if (id) {
+      const url = process.env.CRED_URL + '/credentials/' + id + '/verify';
+
+      var config = {
+        headers: {},
+      };
+      let response_text = null;
+
+      try {
+        const observable = this.httpService.get(url, config);
+        const promise = observable.toPromise();
+        const response = await promise;
+        //console.log(JSON.stringify(response.data));
+        response_text = response.data;
+      } catch (error) {
+        //console.log(e);
+        response_text = { error: error };
+      }
+      if (response_text?.error) {
+        return response.status(400).send({
+          success: false,
+          status: 'cred_schema_api_failed',
+          message: 'Credentials Schema Failed ! Please Try Again.',
+          result: response_text,
+        });
+      } else {
+        return response.status(200).send({
+          success: true,
+          status: 'cred_schema_api_success',
+          message: 'Cred Schema API Success',
+          result: response_text,
+        });
+      }
+    } else {
+      return response.status(400).send({
+        success: false,
+        status: 'invalid_request',
+        message: 'Invalid Request. Not received All Parameters.',
+        result: null,
+      });
+    }
+  }
+
+  //helper function
+  //get convert date and repalce character from string
+  async convertDate(datetime) {
+    if (!datetime) {
+      return '';
+    }
+    let date_string = datetime.substring(0, 10);
+    const datetest = this.moment(date_string, 'DD/MM/YYYY').format(
+      'DD/MM/YYYY',
+    );
+    return datetest;
+  }
+  async replaceChar(replaceString, found, replace) {
+    if (!replaceString) {
+      return '';
+    }
+    const search = found;
+    const replaceWith = replace;
+    const result = replaceString.split(search).join(replaceWith);
+    return result;
+  }
+  //get jwt token information
+  async parseJwt(token) {
+    if (!token) {
+      return [];
+    }
+    const decoded = jwt_decode(token);
+    return [decoded];
+  }
+
+  async generateQRCode(certificateId) {
+    const url = process.env.VERIFICATION_URL + certificateId;
+    //converted into string
+    // let qrcodestring = "";
+
+    let stringData = JSON.stringify(url);
+
+    let qrcodestring = await qr.toDataURL(stringData, function (err, code) {
+      if (code) {
+        qrcodestring = code;
+        console.log(qrcodestring);
+
+        return qrcodestring;
+      } else {
+        return err;
+      }
+    });
+  }
 }
