--- src/sso/constlist/schoollist.ts
@@ -1,401 +1,0 @@
-export const schoolList = [
-  {
-    udiseCode: '09270801003',
-    schoolName: 'JHS ASTI',
-    schoolCategory: 3,
-    schoolManagementCenter: 1,
-    schoolManagementState: 11,
-    schoolType: 1,
-    classFrom: 1,
-    classTo: 5,
-    stateCode: '16',
-    stateName: 'Bihar',
-    districtName: 'GORAKHPUR',
-    blockName: 'BAANSGAON',
-    locationType: 2,
-    headOfSchoolMobile: '89******42',
-    respondentMobile: '88******96',
-    alternateMobile: '',
-    schoolEmail: '',
-  },
-  {
-    udiseCode: '09270801002',
-    schoolName: 'P.S. ASTI-1',
-    schoolCategory: 3,
-    schoolManagementCenter: 1,
-    schoolManagementState: 11,
-    schoolType: 1,
-    classFrom: 1,
-    classTo: 5,
-    stateCode: '16',
-    stateName: 'Bihar',
-    districtName: 'GORAKHPUR',
-    blockName: 'BAANSGAON',
-    locationType: 2,
-    headOfSchoolMobile: '89******42',
-    respondentMobile: '88******96',
-    alternateMobile: '',
-    schoolEmail: '',
-  },
-  {
-    udiseCode: '09270801001',
-    schoolName: 'P.S. ASTI-2',
-    schoolCategory: 3,
-    schoolManagementCenter: 1,
-    schoolManagementState: 11,
-    schoolType: 1,
-    classFrom: 1,
-    classTo: 5,
-    stateCode: '16',
-    stateName: 'Bihar',
-    districtName: 'GORAKHPUR',
-    blockName: 'BAANSGAON',
-    locationType: 2,
-    headOfSchoolMobile: '89******42',
-    respondentMobile: '88******96',
-    alternateMobile: '',
-    schoolEmail: '',
-  },
-  {
-    udiseCode: '09270802002',
-    schoolName: 'JHS. BARAKHEMPUR',
-    schoolCategory: 3,
-    schoolManagementCenter: 1,
-    schoolManagementState: 11,
-    schoolType: 1,
-    classFrom: 1,
-    classTo: 5,
-    stateCode: '16',
-    stateName: 'Bihar',
-    districtName: 'GORAKHPUR',
-    blockName: 'BAANSGAON',
-    locationType: 2,
-    headOfSchoolMobile: '89******42',
-    respondentMobile: '88******96',
-    alternateMobile: '',
-    schoolEmail: '',
-  },
-  {
-    udiseCode: '09270802001',
-    schoolName: 'P.S. BARAKHEMPUR',
-    schoolCategory: 3,
-    schoolManagementCenter: 1,
-    schoolManagementState: 11,
-    schoolType: 1,
-    classFrom: 1,
-    classTo: 5,
-    stateCode: '16',
-    stateName: 'Bihar',
-    districtName: 'GORAKHPUR',
-    blockName: 'BAANSGAON',
-    locationType: 2,
-    headOfSchoolMobile: '89******42',
-    respondentMobile: '88******96',
-    alternateMobile: '',
-    schoolEmail: '',
-  },
-  {
-    udiseCode: '09270801702',
-    schoolName: 'P.S. AMANI GANJ-2',
-    schoolCategory: 3,
-    schoolManagementCenter: 1,
-    schoolManagementState: 11,
-    schoolType: 1,
-    classFrom: 1,
-    classTo: 5,
-    stateCode: '16',
-    stateName: 'Bihar',
-    districtName: 'GORAKHPUR',
-    blockName: 'BAANSGAON',
-    locationType: 2,
-    headOfSchoolMobile: '89******42',
-    respondentMobile: '88******96',
-    alternateMobile: '',
-    schoolEmail: '',
-  },
-  {
-    udiseCode: '09270801705',
-    schoolName: 'VIVEKANAND BAL VIDYA MANDIR',
-    schoolCategory: 3,
-    schoolManagementCenter: 1,
-    schoolManagementState: 11,
-    schoolType: 1,
-    classFrom: 1,
-    classTo: 5,
-    stateCode: '16',
-    stateName: 'Bihar',
-    districtName: 'GORAKHPUR',
-    blockName: 'BAANSGAON',
-    locationType: 2,
-    headOfSchoolMobile: '89******42',
-    respondentMobile: '88******96',
-    alternateMobile: '',
-    schoolEmail: '',
-  },
-  {
-    udiseCode: '09270800701',
-    schoolName: 'P.S. ARAMBA',
-    schoolCategory: 3,
-    schoolManagementCenter: 1,
-    schoolManagementState: 11,
-    schoolType: 1,
-    classFrom: 1,
-    classTo: 5,
-    stateCode: '16',
-    stateName: 'Bihar',
-    districtName: 'GORAKHPUR',
-    blockName: 'BAANSGAON',
-    locationType: 2,
-    headOfSchoolMobile: '89******42',
-    respondentMobile: '88******96',
-    alternateMobile: '',
-    schoolEmail: '',
-  },
-  {
-    udiseCode: '09270801704',
-    schoolName: 'JHS AMANIGANJ',
-    schoolCategory: 3,
-    schoolManagementCenter: 1,
-    schoolManagementState: 11,
-    schoolType: 1,
-    classFrom: 1,
-    classTo: 5,
-    stateCode: '16',
-    stateName: 'Bihar',
-    districtName: 'GORAKHPUR',
-    blockName: 'BAANSGAON',
-    locationType: 2,
-    headOfSchoolMobile: '89******42',
-    respondentMobile: '88******96',
-    alternateMobile: '',
-    schoolEmail: '',
-  },
-  {
-    udiseCode: '99010200777',
-    schoolName: 'New PUBLIC SCHOOL',
-    schoolCategory: 3,
-    schoolManagementCenter: 1,
-    schoolManagementState: 11,
-    schoolType: 1,
-    classFrom: 1,
-    classTo: 5,
-    stateCode: '16',
-    stateName: 'Bihar',
-    districtName: 'GORAKHPUR',
-    blockName: 'BAANSGAON',
-    locationType: 2,
-    headOfSchoolMobile: '89******42',
-    respondentMobile: '88******96',
-    alternateMobile: '',
-    schoolEmail: '',
-  },
-  {
-    udiseCode: '99010200000',
-    schoolName: 'Old PUBLIC SCHOOL',
-    schoolCategory: 3,
-    schoolManagementCenter: 1,
-    schoolManagementState: 11,
-    schoolType: 1,
-    classFrom: 1,
-    classTo: 5,
-    stateCode: '16',
-    stateName: 'Bihar',
-    districtName: 'GORAKHPUR',
-    blockName: 'BAANSGAON',
-    locationType: 2,
-    headOfSchoolMobile: '89******42',
-    respondentMobile: '88******96',
-    alternateMobile: '',
-    schoolEmail: '',
-  },
-  {
-    udiseCode: '99010200104',
-    schoolName: 'ALHIRA Co-Educational PUBLIC SCHOOL',
-    schoolCategory: 3,
-    schoolManagementCenter: 1,
-    schoolManagementState: 11,
-    schoolType: 1,
-    classFrom: 1,
-    classTo: 5,
-    stateCode: '16',
-    stateName: 'Bihar',
-    districtName: 'GORAKHPUR',
-    blockName: 'BAANSGAON',
-    locationType: 2,
-    headOfSchoolMobile: '89******42',
-    respondentMobile: '88******96',
-    alternateMobile: '',
-    schoolEmail: '',
-  },
-  {
-    udiseCode: '99010300108',
-    schoolName: 'Boys Higher Secodary School',
-    schoolCategory: 1,
-    schoolManagementCenter: 1,
-    schoolManagementState: 11,
-    schoolType: 3,
-    classFrom: 1,
-    classTo: 5,
-    stateCode: '16',
-    stateName: 'UTTAR PRADESH',
-    districtName: 'GORAKHPUR',
-    blockName: 'BAANSGAON',
-    locationType: 2,
-    headOfSchoolMobile: '89******42',
-    respondentMobile: '88******96',
-    alternateMobile: '',
-    schoolEmail: '',
-  },
-  {
-    udiseCode: '09580413510',
-    schoolName: 'A.S. MANDIR',
-    schoolCategory: 1,
-    schoolManagementCenter: 1,
-    schoolManagementState: 11,
-    schoolType: 3,
-    classFrom: 1,
-    classTo: 5,
-    stateCode: '16',
-    stateName: 'UTTAR PRADESH',
-    districtName: 'GORAKHPUR',
-    blockName: 'BAANSGAON',
-    locationType: 2,
-    headOfSchoolMobile: '89******42',
-    respondentMobile: '88******96',
-    alternateMobile: '',
-    schoolEmail: '',
-  },
-  {
-    udiseCode: '09580413507',
-    schoolName: 'B.S.S. PS DONKHAR',
-    schoolCategory: 1,
-    schoolManagementCenter: 1,
-    schoolManagementState: 11,
-    schoolType: 3,
-    classFrom: 1,
-    classTo: 5,
-    stateCode: '16',
-    stateName: 'UTTAR PRADESH',
-    districtName: 'GORAKHPUR',
-    blockName: 'BAANSGAON',
-    locationType: 2,
-    headOfSchoolMobile: '89******42',
-    respondentMobile: '88******96',
-    alternateMobile: '',
-    schoolEmail: '',
-  },
-  {
-    udiseCode: '09580413502',
-    schoolName: 'CENTRAL PUBLIC ACEDEMY',
-    schoolCategory: 1,
-    schoolManagementCenter: 1,
-    schoolManagementState: 11,
-    schoolType: 3,
-    classFrom: 1,
-    classTo: 5,
-    stateCode: '16',
-    stateName: 'UTTAR PRADESH',
-    districtName: 'GORAKHPUR',
-    blockName: 'BAANSGAON',
-    locationType: 2,
-    headOfSchoolMobile: '89******42',
-    respondentMobile: '88******96',
-    alternateMobile: '',
-    schoolEmail: '',
-  },
-  {
-    udiseCode: '09580413509',
-    schoolName: 'D.A.B. JHS SCHOOL',
-    schoolCategory: 1,
-    schoolManagementCenter: 1,
-    schoolManagementState: 11,
-    schoolType: 3,
-    classFrom: 1,
-    classTo: 5,
-    stateCode: '16',
-    stateName: 'UTTAR PRADESH',
-    districtName: 'GORAKHPUR',
-    blockName: 'BAANSGAON',
-    locationType: 2,
-    headOfSchoolMobile: '89******42',
-    respondentMobile: '88******96',
-    alternateMobile: '',
-    schoolEmail: '',
-  },
-  {
-    udiseCode: '09580413512',
-    schoolName: 'LATE Y.S. JHS SCHOOL BALUA',
-    schoolCategory: 1,
-    schoolManagementCenter: 1,
-    schoolManagementState: 11,
-    schoolType: 3,
-    classFrom: 1,
-    classTo: 5,
-    stateCode: '16',
-    stateName: 'UTTAR PRADESH',
-    districtName: 'GORAKHPUR',
-    blockName: 'BAANSGAON',
-    locationType: 2,
-    headOfSchoolMobile: '89******42',
-    respondentMobile: '88******96',
-    alternateMobile: '',
-    schoolEmail: '',
-  },
-  {
-    udiseCode: '09580413508',
-    schoolName: 'P.J.L.N. I. COLLEGE',
-    schoolCategory: 1,
-    schoolManagementCenter: 1,
-    schoolManagementState: 11,
-    schoolType: 3,
-    classFrom: 1,
-    classTo: 5,
-    stateCode: '16',
-    stateName: 'UTTAR PRADESH',
-    districtName: 'GORAKHPUR',
-    blockName: 'BAANSGAON',
-    locationType: 2,
-    headOfSchoolMobile: '89******42',
-    respondentMobile: '88******96',
-    alternateMobile: '',
-    schoolEmail: '',
-  },
-  {
-    udiseCode: '09580413503',
-    schoolName: 'P.M.V.BAANSGAON',
-    schoolCategory: 1,
-    schoolManagementCenter: 1,
-    schoolManagementState: 11,
-    schoolType: 3,
-    classFrom: 1,
-    classTo: 5,
-    stateCode: '16',
-    stateName: 'UTTAR PRADESH',
-    districtName: 'GORAKHPUR',
-    blockName: 'BAANSGAON',
-    locationType: 2,
-    headOfSchoolMobile: '89******42',
-    respondentMobile: '88******96',
-    alternateMobile: '',
-    schoolEmail: '',
-  },
-  {
-    udiseCode: '09580413513',
-    schoolName: 'P.S.NASIRPUR',
-    schoolCategory: 1,
-    schoolManagementCenter: 1,
-    schoolManagementState: 11,
-    schoolType: 3,
-    classFrom: 1,
-    classTo: 5,
-    stateCode: '16',
-    stateName: 'UTTAR PRADESH',
-    districtName: 'GORAKHPUR',
-    blockName: 'BAANSGAON',
-    locationType: 2,
-    headOfSchoolMobile: '89******42',
-    respondentMobile: '88******96',
-    alternateMobile: '',
-    schoolEmail: '',
-  },
-];
