--- src/sso/dto/user-dto.ts
@@ -1,10 +1,0 @@
-export class UserDto {
-    
-    aadhaarId: string;
-    studentName: string;
-    schoolName: string;
-    schoolId: string;
-    studentId: string
-    phoneNo: string
-
-}
